fix: naturally ignore callback dependencies
- ℹ️ Hide whitespace changes for this commit.
