+++ utils/compute-dependencies-outside-functions-by-sorting-node.ts
@@ -1,14 +1,11 @@
 import type { TSESTree } from '@typescript-eslint/types'
 import type { TSESLint } from '@typescript-eslint/utils'
 
-import { AST_NODE_TYPES } from '@typescript-eslint/utils'
-
 import type { ShouldIgnoreIdentifierComputer } from './compute-dependencies-by-sorting-node'
 import type { SortingNodeWithDependencies } from './sort-nodes-by-dependencies'
 
 import { computeDependenciesBySortingNode } from './compute-dependencies-by-sorting-node'
-import { computeParentNodesWithTypes } from './compute-parent-nodes-with-types'
-import { isNodeImmediatelyCalled } from './is-node-immediately-called'
+import { isNodeInsideDeferredFunction } from './is-node-inside-deferred-function'
 
 export function computeDependenciesOutsideFunctionsBySortingNode<
   Node extends TSESTree.Node,
@@ -27,20 +24,10 @@ export function computeDependenciesOutsideFunctionsBySortingNode<
   })
 
   function buildShouldIgnoreIdentifierComputer(): ShouldIgnoreIdentifierComputer<T> {
-    return ({ referencingSortingNode, identifier }) => {
-      let functionParentNodes = computeParentNodesWithTypes({
-        allowedTypes: [
-          AST_NODE_TYPES.FunctionExpression,
-          AST_NODE_TYPES.ArrowFunctionExpression,
-        ],
+    return ({ referencingSortingNode, identifier }) =>
+      isNodeInsideDeferredFunction({
         maxParent: referencingSortingNode.node,
-        consecutiveOnly: false,
         node: identifier,
       })
-
-      return functionParentNodes.some(
-        functionParentNode => !isNodeImmediatelyCalled(functionParentNode),
-      )
-    }
   }
 }
